refactor: - 기존 프로젝트 레이어드 아키텍쳐에 맞게 리팩토링 - [x]  PlateApprearance를 퍼시스턴스로 분리 - [x]  batting 서비스 클래스 - [x]  기존 /game/batting 컨트롤러에 서비스 연결
