+++ src/main/java/com/hyunec/cosmicbaseballinit/domain/baseball/web/GameController.java
@@ -1,0 +1,23 @@
+package com.hyunec.cosmicbaseballinit.domain.baseball.web;
+
+import com.hyunec.cosmicbaseballinit.domain.baseball.model.BattingResult;
+import com.hyunec.cosmicbaseballinit.domain.baseball.service.GameService;
+import lombok.RequiredArgsConstructor;
+import lombok.extern.slf4j.Slf4j;
+import org.springframework.http.ResponseEntity;
+import org.springframework.web.bind.annotation.GetMapping;
+import org.springframework.web.bind.annotation.RestController;
+
+@Slf4j
+@RequiredArgsConstructor
+@RestController
+public class GameController {
+
+    private final GameService gameService;
+
+    @GetMapping("/game/batting")
+    public ResponseEntity<BattingResult> batting() {
+        BattingResult result = gameService.batting();
+        return ResponseEntity.ok().body(result);
+    }
+}
